Folding: move EvalLeaf into sub module
It makes the lib.rs cleaner
